pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "argus"
version = "0.1.0"
description = "A smart contract security tool"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Caleb Chin", email = "cchin@cs.toronto.edu" },
    { name = "Edward Liu", email = "edwardy.liu@mail.utoronto.ca" },
    { name = "Jonathan Wen", email = "jon.wen@mail.utoronto.ca" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "google-genai>=1.0.0",
    "anthropic>=0.18.0",
    "mcp>=1.22.0",
    "click>=8.0.0",
    "docker>=7.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.28.1",
]
dev = [
    "argus[test]",
    "setuptools>=68.0",
    "wheel",
    "build",
    "pylint>=4.0.3",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0",
]

[project.scripts]
argus = "argus.core.cli:main"

[project.urls]
Homepage = "https://github.com/calebchin/argus"
Repository = "https://github.com/calebchin/argus"

[project.entry-points."argus.llm.providers"]
anthropic = "argus.llm.providers:AnthropicProviderPlugin"
gemini = "argus.llm.providers:GeminiProviderPlugin"

[project.entry-points."argus.mcp.tools"]
filesystem = "argus.server.tools:FilesystemToolPlugin"
mythril = "argus.server.tools:MythrilToolPlugin"
slither = "argus.server.tools:SlitherToolPlugin"

[project.entry-points."argus.mcp.resources"]
filesystem = "argus.server.resources:FilesystemResourcePlugin"

[project.entry-points."argus.mcp.prompts"]
# reserved for future built-in prompt plugins

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=argus --cov-report=term-missing"
asyncio_mode = "auto"
log_cli_format = "%(message)s"
markers = [
    "integration: marks tests as integration tests (require Docker, network access)",
    "unit: marks tests as unit tests (fast, no external dependencies)",
]
